==> Mathew/code/01_load_raw_data.ipynb <==
## Mathew: load unprocessed Cell Ranger and salmon data

import numpy as np
import pandas as pd
import os
import scanpy as sc
import anndata as ad
from scipy.sparse import csr_matrix
from pyroe import load_fry
import matplotlib.pyplot as plt
from matplotlib.colors import ListedColormap
from matplotlib_venn import venn2
from venny4py.venny4py import *
import scvelo as scv
import re
import louvain
from skmisc.loess import loess
from skbio.diversity.alpha import simpson, gini_index
import warnings
import sys
import ensembl_rest
import time
from Ensembl_converter import EnsemblConverter

## Cell Ranger

### load expression matrices

# load metadata
meta = pd.read_csv("../data/metadata.csv", index_col=0)
meta

adatas = []
path = "../cellranger"

for _, folders, _ in os.walk(path):
    for folder in folders:
        if folder.startswith("Sample"):

            # get adata
            adata = sc.read_10x_mtx(path + "/" + folder)
            
            # add metadata
            for col_name, col_data in meta.loc[folder].items():
                adata.obs[col_name] = col_data
    
            # add ID
            adata.obs["id"] = folder

            # rename obs names
            adata.obs_names = [bc[:-2] + "_" + folder for bc in adata.obs_names]
    
            # append to list
            adatas.append(adata)

# merge all samples
cellranger = ad.concat(adatas)
cellranger

# save
cellranger.write_h5ad("../data/mathew_cellranger_raw.h5ad")

## Salmon

### load expression matrices

adatas = []
path = "../salmon"

for _, folders, _ in os.walk(path):
    for folder in folders:

        if folder.startswith("SampleID"):
        
            # get adata
            adata = load_fry(path + "/" + folder + "/_quant_res", output_format = "all")
            
            # add metadata
            for col_name, col_data in meta.loc[folder].items():
                adata.obs[col_name] = col_data
    
            # add ID
            adata.obs["id"] = folder

            # rename obs names
            adata.obs_names = [bc + "_" + folder for bc in adata.obs_names]
    
            # append to list
            adatas.append(adata)

# merge all samples
salmon = ad.concat(adatas)
salmon

# rename genes from ensembl ID to gene ID
e2n_path = "../../splici_references/mm10-2020-A_geneid_to_name.txt"
e2n = dict([ l.rstrip().split() for l in open(e2n_path).readlines()])
salmon.var_names = [e2n[e] for e in salmon.var_names]

### add spliced, unspliced and ambiguous counts

# load alevin-fry output files using mode = raw
adatas_usa = []

for _, folders, _ in os.walk(path):
    for folder in folders:

        if folder.startswith("SampleID"):
        
            # get adata
            adata = load_fry(path + "/" + folder + "/_quant_res", output_format = "raw")
            
            # add metadata
            for col_name, col_data in meta.loc[folder].items():
                adata.obs[col_name] = col_data
    
            # add ID
            adata.obs["id"] = folder

            # rename obs names
            adata.obs_names = [bc + "_" + folder for bc in adata.obs_names]
    
            # append to list
            adatas_usa.append(adata)

# merge all samples
usa = ad.concat(adatas_usa)
usa

# convert gene IDs to gene names
usa.var_names = [e2n[e] for e in usa.var_names]
usa.var_names_make_unique()

# get same bc and cell ordering as in adata object
usa = usa[salmon.obs_names, salmon.var_names]

# add spliced, unspliced, ambiguous layers to hb6
salmon.layers["spliced"] = usa.layers["spliced"]
salmon.layers["unspliced"] = usa.layers["unspliced"]
salmon.layers["ambiguous"] = usa.layers["ambiguous"]

# save
salmon.write_h5ad("../data/mathew_salmon_raw.h5ad")
