==> Mathew/code/05_calculate_velocities.ipynb <==
## Mathew: calculate velocities

import numpy as np
import pandas as pd
import scipy.stats as stats
import scanpy as sc
import anndata as ad
import matplotlib.pyplot as plt
import seaborn as sns
from matplotlib.colors import ListedColormap
import scvelo as scv

### colormap for UMAP

# create colormap for UMAP

# list of HEX colors
hex_colors = ["#CD665F", "#BF781C", "#AB871D", "#92941D", "#77A119", "#57AA17", "#64B15A", "#5CB488",
              "#61AFB4", "#5AA6DC", "#5A91F8", "#7D7AFA", "#A760FA", "#C54FE4", "#CB4EC0", "#D3558E"]

# convert HEX colors to RGB
rgb_colors = [(int(color[1:3], 16)/255, int(color[3:5], 16)/255, int(color[5:7], 16)/255) for color in hex_colors]

# create colormap
cmap = ListedColormap(rgb_colors)

# create dictionary to assign colors to clusters
clusters = ["C1", "C2", "C3", "C4", "C5", "C6", "C7", "C8", "C9", "C10", "C11", "C12", "C13", "C14", "C15", "C16"]
umap_colors = {}
i = 0
for cluster in clusters:
    umap_colors[cluster] = cmap.colors[i]
    i+=1

### read in data

cellranger/velocyto files

# adata file from cellranger
cellranger = sc.read_h5ad("../data/mathew_cellranger.h5ad")

# loom files from velocyto

# get all samples
path = "../velocyto"
samples = ["SampleID_1_27feb19", "SampleID_2_29apr19", "SampleID_2_27feb19", "SampleID_4_2apr19", "SampleID_4_26apr19",
           "SampleID_3_27feb19", "SampleID_4_1apr19", "SampleID_1_14mar19", "SampleID_5_2apr19", "SampleID_5_26apr19",
           "SampleID_2_14mar19", "SampleID_2_2apr19", "SampleID_3_26apr19", "SampleID_3_29apr19", "SampleID_1_26apr19",
           "SampleID_1_29apr19", "SampleID_5_14mar19", "SampleID_1_2apr19", "SampleID_2_26apr19", "SampleID_51_24feb20",
           "SampleID_52_24feb20", "SampleID_53_24feb20", "SampleID_54_24feb20", "SampleID_55_24feb20"]

# initialize loom data with first file
ldata = scv.read(path + "/" + samples[0] + ".loom")
ldata.var_names_make_unique()
ldata.obs['dataset'] = folders[0]

# append all other loom files
for samples in samples[1:]:
    loom = scv.read(path + "/" + sample + ".loom")
    loom.var_names_make_unique()
    loom.obs['dataset'] = folder
    
    ldata = ldata.concatenate(loom, join='outer', index_unique=None)

# rename barcodes
ldata.obs_names = [re.split(':|x', x)[1] + '_' + ldata.obs['dataset'][x] for x in ldata.obs_names]

# merge adata with loom files
velocyto = scv.utils.merge(cellranger, ldata)

# save
velocyto.write_h5ad(filename="../data/mathew_velocyto.h5ad")

alevin-fry file

alevin = sc.read_h5ad("../data/mathew_salmon.h5ad")

### compute velocities

# preprocessing
scv.pp.filter_genes(velocyto)
scv.pp.normalize_per_cell(velocyto)

scv.pp.filter_genes(alevin)
scv.pp.normalize_per_cell(alevin)

# ensure that highly variable genes are the same
assert (velocyto.var["highly_variable"].sort_index()).equals(alevin.var["highly_variable"].sort_index()), "computing velocities on different genes"

# compute neighborhood graph and moments for velocity estimation
scv.pp.moments(velocyto)
scv.pp.moments(alevin)

# separate into 3 objects for 3 modes (deterministic, stochastic, dynamical)
velocyto_det = velocyto.copy()
velocyto_sto = velocyto.copy()
velocyto_dyn = velocyto.copy()

alevin_det = alevin.copy()
alevin_sto = alevin.copy()
alevin_dyn = alevin.copy()

#### model = deterministic

# calculate velocities: deterministic (= steady state = velocyto)
scv.tl.velocity(velocyto_det, mode='deterministic', filter_genes=False)
scv.tl.velocity_graph(velocyto_det)

scv.tl.velocity(alevin_det, mode='deterministic', filter_genes=False)
scv.tl.velocity_graph(alevin_det)

# embedding in UMAP
scv.pl.velocity_embedding_grid(velocyto_det, basis='umap', color="cluster", arrow_length = 2, palette=umap_colors,
                               legend_loc = 'right_margin', xlabel = "UMAP 1", ylabel = "UMAP 2")

scv.pl.velocity_embedding_grid(alevin_det, basis='umap', color="cluster", arrow_length = 2, palette=umap_colors,
                               legend_loc = 'right_margin', xlabel = "UMAP 1", ylabel = "UMAP 2")

#### model = stochastic

# calculate velocities: stochastic 
scv.tl.velocity(velocyto_sto, mode='stochastic', filter_genes=False)
scv.tl.velocity_graph(velocyto_sto)

scv.tl.velocity(alevin_sto, mode='stochastic', filter_genes=False)
scv.tl.velocity_graph(alevin_sto)

# embedding in UMAP
scv.pl.velocity_embedding_grid(velocyto_sto, basis='umap', color="cluster", arrow_length = 2, palette=umap_colors,
                               legend_loc = 'right_margin', xlabel = "UMAP 1", ylabel = "UMAP 2")

scv.pl.velocity_embedding_grid(alevin_sto, basis='umap', color="cluster", arrow_length = 2, palette=umap_colors,
                               legend_loc = 'right_margin', xlabel = "UMAP 1", ylabel = "UMAP 2")

#### model = dynamical

# calculate velocities: dynamical 
scv.tl.recover_dynamics(velocyto_dyn)
scv.tl.velocity(velocyto_dyn, mode="dynamical", filter_genes=False)

scv.tl.recover_dynamics(alevin_dyn)
scv.tl.velocity(alevin_dyn, mode="dynamical", filter_genes=False)

# velocity graph
scv.tl.velocity_graph(velocyto_dyn)

scv.tl.velocity_graph(alevin_dyn)

# embedding in UMAP
scv.pl.velocity_embedding_grid(velocyto_dyn, basis='umap', color="cluster", arrow_length = 2, palette=umap_colors,
                               legend_loc = 'right_margin', xlabel = "UMAP 1", ylabel = "UMAP 2")

scv.pl.velocity_embedding_grid(alevin_dyn, basis='umap', color="cluster", arrow_length = 2, palette=umap_colors,
                               legend_loc = 'right_margin', xlabel = "UMAP 1", ylabel = "UMAP 2")

### save

velocyto_sto.write_h5ad(filename="../data/mathew_velocyto_sto.h5ad")
alevin_sto.write_h5ad(filename="../data/mathew_alevin_sto.h5ad")
